add common folder && move some types and consants over to common folder
