+++ src/features/pastelStatistics/components/chart/EChartsMultiLineChart.tsx
@@ -5,69 +5,34 @@ import * as echarts from 'echarts'
 import { CSVLink } from 'react-csv'
 import { Data } from 'react-csv/components/CommonPropTypes'
 import { saveAs } from 'file-saver'
+import { makeDownloadFileName } from '../../utils/PastelStatisticsLib'
+import { LineChartProps, TThemeColor } from '../../common/types'
+import { csvHeaders, themes } from '../../common/constants'
+
 import styles from './LineChart.module.css'
-import { makeDownloadFileName, TPeriod } from '../../utils/PastelStatisticsLib'
-
-type TThemeColor = {
-  name: string
-  backgroundColor: string
-  lineColor?: string
-  splitLineColor: string
-  color: string
-  stack?: string
-  smooth?: boolean
-}
-
-type LineChartProps = {
-  dataX?: string[]
-  dataY1?: number[]
-  dataY2?: number[]
-  title?: string
-  handlePeriodFilterChange?: (period: TPeriod) => void
-  handleBgColorChange: (color: string) => void
-}
 
 export const EChartsMultiLineChart = (props: LineChartProps): JSX.Element => {
+  const {
+    dataX,
+    dataY1,
+    dataY2,
+    info,
+    periods,
+    title,
+    handlePeriodFilterChange,
+    handleBgColorChange,
+  } = props
+  const downloadRef = useRef(null)
+  const [csvData] = useState<string | Data>('')
   const [currentTheme, setCurrentTheme] = useState<TThemeColor | null>()
   const [eChartRef, setEChartRef] = useState<ReactECharts | null>()
   const [eChartInstance, setEChartInstance] = useState<echarts.ECharts>()
   const [selectedPeriodButton, setSelectedPeriodButton] = useState(0)
   const [selectedThemeButton, setSelectedThemeButton] = useState(0)
-  const {
-    dataX,
-    dataY1,
-    dataY2,
-    title,
-    handlePeriodFilterChange,
-    handleBgColorChange,
-  } = props
   const [minY1, setMinY1] = useState(0)
   const [minY2, setMinY2] = useState(0)
   const [maxY1, setMaxY1] = useState(0)
   const [maxY2, setMaxY2] = useState(0)
-  const periods: TPeriod[] = ['2h', '2d', '4d', 'all']
-  const themes = [
-    {
-      name: 'theme1',
-      backgroundColor: '#100c2a',
-      splitLineColor: '#202021',
-      color: '#abaac1',
-    },
-    {
-      name: 'theme2',
-      backgroundColor: '#FFF7C6',
-      stack: 'confidence-band',
-      splitLineColor: '#C7C4CC',
-      smooth: true,
-      color: '#100c2a',
-    },
-    {
-      name: 'theme3',
-      backgroundColor: '#FFF',
-      splitLineColor: '#EEE',
-      color: '#202021',
-    },
-  ]
 
   useEffect(() => {
     const chartInstance = eChartRef?.getEchartsInstance()
@@ -89,12 +54,6 @@ export const EChartsMultiLineChart = (props: LineChartProps): JSX.Element => {
     }
   }, [dataY1, dataY2])
 
-  const [csvData] = useState<string | Data>('')
-  const csvHeaders = [
-    { label: 'Value', key: 'value' },
-    { label: 'Created time', key: 'time' },
-  ]
-  const downloadRef = useRef(null)
   const options = {
     grid: {
       top: 50,
@@ -189,7 +148,10 @@ export const EChartsMultiLineChart = (props: LineChartProps): JSX.Element => {
         .toBlob(eChartRef.ele)
         .then(function (blob: Blob | null) {
           if (blob) {
-            saveAs(blob, makeDownloadFileName(title ?? '') + '.png')
+            saveAs(
+              blob,
+              makeDownloadFileName(info.currencyName, title ?? '') + '.png',
+            )
           }
         })
         .catch(function (error) {
@@ -284,23 +246,22 @@ export const EChartsMultiLineChart = (props: LineChartProps): JSX.Element => {
         </div>
         <div className={styles.periodSelect}>
           <span style={{ color: currentTheme?.color }}>period: </span>
-          {periods &&
-            periods.map((period, index) => (
-              <button
-                className={`${getActivePriodButtonStyle(index)} 
+          {periods.map((period, index) => (
+            <button
+              className={`${getActivePriodButtonStyle(index)} 
                   ${styles.filterButton}`}
-                onClick={() => {
-                  setSelectedPeriodButton(index)
-                  if (handlePeriodFilterChange) {
-                    handlePeriodFilterChange(period)
-                  }
-                }}
-                type='button'
-                key={`button-filter-${period}`}
-              >
-                {period}
-              </button>
-            ))}
+              onClick={() => {
+                setSelectedPeriodButton(index)
+                if (handlePeriodFilterChange) {
+                  handlePeriodFilterChange(period)
+                }
+              }}
+              type='button'
+              key={`button-filter-${period}`}
+            >
+              {period}
+            </button>
+          ))}
         </div>
       </div>
       <div className={styles.lineChartWrap}>
@@ -338,7 +299,9 @@ export const EChartsMultiLineChart = (props: LineChartProps): JSX.Element => {
           </button>
           <CSVLink
             data={csvData}
-            filename={makeDownloadFileName(title ?? '') + '.csv'}
+            filename={
+              makeDownloadFileName(info.currencyName, title ?? '') + '.csv'
+            }
             headers={csvHeaders}
             separator={';'}
             ref={downloadRef}
