--- src/features/pastelStatistics/constants/common.ts
@@ -1,15 +1,0 @@
-import Difficulty from '../assets/images/difficulty.jpg'
-import PSLPrice from '../assets/images/pslprice.jpg'
-
-export const pastelChartFields = [
-  {
-    name: 'Difficulty',
-    routeName: '/statistics/difficultyovertime',
-    backgroundImage: Difficulty,
-  },
-  {
-    name: 'PSL Price',
-    routeName: '/statistics/priceovertime',
-    backgroundImage: PSLPrice,
-  },
-]
